requirements.txt
langchain==0.1.0
python-dotenv==1.0.0
openai==1.12.0
pypdf==4.0.1
chromadb==0.4.22
tiktoken==0.4.0
